Refactor buttonbar into its own component
Refactor the buttonbar and its associated window resize logic into its
own component.

Closes #465
